\name{AQM-package}
\alias{AQM-package}
\alias{AQM}
\docType{package}
\title{
AQM - Advanced Quantitative Methods for Political Science
}
\description{
All estimation procedure of the lecture Advanced Quantitative Methods University Mannheim (Gschwend/Stötzer)
}
\details{
\tabular{ll}{
Package: \tab AQM\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-03-06\cr
License: \tab What license is it under?\cr
}
The package will be updated after each week of the lecture to include new estimation procedures
}
\author{
Lukas Stötzer and Students of the AQM class
lstoetze@mail.uni-mannheim.de
}
